+++ test/acceptance/w3c-test-suites/bind_test.exs
@@ -12,12 +12,6 @@ defmodule SPARQL.W3C.TestSuite.BindTest do
 
   TestSuite.test_cases(@test_suite, MF.QueryEvaluationTest)
   |> Enum.each(fn test_case ->
-      ["bind07"]
-      |> Enum.each(fn test_subject ->
-        if test_case.subject |> to_string() |> String.ends_with?(test_subject),
-           do: @tag skip: "TODO: UNION"
-      end)
-
        @tag test_case: test_case
        test TestSuite.test_title(test_case), %{test_case: test_case} do
          assert_query_evaluation_case_result(test_case, @manifest_graph)