+++ test/acceptance/w3c-test-suites/reduced_test.exs
@@ -18,12 +18,6 @@ defmodule SPARQL.W3C.TestSuite.ReducedTest do
               do: @tag skip: "TODO: Differentiate simple literals from string literals?"
           end)
 
-       ["reduced-1"]
-       |> Enum.each(fn test_subject ->
-            if test_case.subject |> to_string() |> String.ends_with?(test_subject),
-              do: @tag skip: "TODO: UNION"
-          end)
-
        @tag test_case: test_case
        test TestSuite.test_title(test_case), %{test_case: test_case} do
          assert_query_evaluation_case_result(test_case, @manifest_graph)